Move Trie to base-services
